Adding s2s api authentication
