+++ src/main/java/uk/gov/hmcts/reform/camunda/bpm/config/GroupConfig.java
@@ -1,11 +1,14 @@
 package uk.gov.hmcts.reform.camunda.bpm.config;
 
+import java.util.List;
+
 public class GroupConfig {
 
     private String adGroupId;
     private String tenantId;
     private String groupId;
     private String accessControl;
+    private List<String> s2sServiceNames;
 
     public String getAdGroupId() {
         return adGroupId;
@@ -39,4 +42,11 @@ public class GroupConfig {
         this.accessControl = accessControl;
     }
 
+    public List<String> getS2sServiceNames() {
+        return s2sServiceNames;
+    }
+
+    public void setS2sServiceNames(List<String> s2sServiceNames) {
+        this.s2sServiceNames = s2sServiceNames;
+    }
 }
