--- src/main/java/uk/gov/hmcts/reform/camunda/bpm/config/WebSecurityAnonymousConfig.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.camunda.bpm.config;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@Configuration
-@ConditionalOnProperty(prefix = "spring.security", name = "enabled", havingValue = "false")
-@EnableWebSecurity
-public class WebSecurityAnonymousConfig extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.csrf().disable()
-            .authorizeRequests().anyRequest().anonymous().and().httpBasic().disable();
-    }
-}
