+++ src/main/resources/application.yaml
@@ -20,6 +20,12 @@ spring:
     username: ${CAMUNDA_DB_USER_NAME:camunda}
     password: ${CAMUNDA_DB_PASSWORD:camunda}
 camunda:
+  api:
+    auth:
+      enabled: true
+  ui: 
+    auth:
+      enabled: false
   bpm:
     authorization:
       enabled: true
@@ -54,6 +60,8 @@ camundaGroups:
     tenantId: civil-unspecified
     groupId: civil-unspecified
     accessControl: user
+    s2sServiceNames:
+      - unspec-service
 
 camundaAdminGroupId: c36eaede-a0ae-4967-8fed-0a02960b1370
 camundaAccess:
@@ -70,7 +78,9 @@ server:
   servlet:
     context-path: /
   port: 8999
-
+idam:
+  s2s-auth:
+    url: ${S2S_URL:http://localhost:4552}
 ---
 spring:
   profiles: springauth
@@ -89,3 +99,7 @@ spring:
         provider:
           azure:
             issuer-uri: ${ISSUER_URI:https://sts.windows.net/531ff96d-0ae9-462a-8d2d-bec7c0b42082/}
+camunda:
+  ui:
+    auth:
+      enabled: true
